Publisher/requirement.txt
kafka-python
pandas
numpy
requests
Flask